About to embark on figuring out how to capture the offset of the sub-parser's completion
