+++ src/simple.rs
@@ -543,3 +543,283 @@ named!(pub indirect_reference<&[u8],PdfObject>,
         )
     )
 );
+
+
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn indirect_reference_test() {
+        assert_eq!(PdfObject::IndirectReference { number: 95, generation: 1 },
+                   indirect_reference(b"95 1 R ".as_bytes()).to_result().unwrap());
+        assert_eq!(PdfObject::IndirectReference { number: 95, generation: 100 },
+                   indirect_reference(b"95 100 R ".as_bytes()).to_result().unwrap());
+        assert_eq!(Err(nom::Err::Code(nom::ErrorKind::RegexpFind)), indirect_reference(b"09 1 R".as_bytes()).to_result());
+    }
+
+    #[test]
+    fn null_test() {
+        match null_object(b"null").to_result().unwrap() {
+            PdfObject::Null => {},
+            _ => {
+                assert_eq!(3, 0);
+            }
+        }
+        match null_object(b"nul") {
+            Incomplete(_) => {}
+            _ => {
+                assert_eq!(1, 0);
+            }
+        }
+        match null_object(b"mul") {
+            Error(_) => {}
+            _ => {
+                assert_eq!(2, 0);
+            }
+        }
+    }
+
+    #[test]
+    fn boolean_test() {
+        assert_eq!(PdfObject::Boolean(true), boolean(b"true ").to_result().unwrap());
+        assert_eq!(PdfObject::Boolean(false), boolean(b"false").to_result().unwrap());
+    }
+
+    #[test]
+    fn magic_test() {
+        assert_eq!(PdfVersion::Known { ver: b"1.0".to_vec() }, pdf_magic(b"%PDF-1.0\r\n").to_result().unwrap());
+        assert_eq!(nom::Err::Position(nom::ErrorKind::Tag, &[98u8, 108u8, 97u8, 104u8][..]),
+                   pdf_magic(b"blah").to_result().unwrap_err());
+        assert_eq!(nom::Err::Code(nom::ErrorKind::RegexpFind),
+                   pdf_magic(b"%PDF-3.0\r").to_result().unwrap_err());
+    }
+
+    #[test]
+    fn linendings_by_macro_test() {
+        assert_eq!(b"\r".as_bytes(), pdf_line_ending_by_macro(b"\rdd".as_bytes()).to_result().unwrap());
+        assert_eq!(b"\r\n".as_bytes(), pdf_line_ending_by_macro(b"\r\ndd".as_bytes()).to_result().unwrap());
+        assert_eq!(b"\n".as_bytes(), pdf_line_ending_by_macro(b"\ndd".as_bytes()).to_result().unwrap());
+    }
+
+    #[test]
+    fn comments_test() {
+        match comment(b"%hiya\n").to_result().unwrap() {
+            PdfObject::Comment(v) => {
+                assert_eq!(v, "hiya".as_bytes());
+            },
+            _ => { assert_eq!(6, 0); }
+        }
+        match comment("%なななな\n".as_bytes()).to_result().unwrap() {
+            PdfObject::Comment(v) => {
+                assert_eq!(v, "なななな".as_bytes());
+            },
+            _ => { assert_eq!(6, 0); }
+        }
+    }
+
+    #[test]
+    fn header_test() {
+        assert_eq!(nom::IResult::Done(b" ".as_bytes(), PdfVersion::Known { ver: b"1.0".to_vec() }),
+                   pdf_header(b"%PDF-1.0\r ".as_bytes()));
+        assert_eq!(nom::IResult::Done(b" ".as_bytes(), PdfVersion::Known { ver: b"2.0".to_vec() }),
+                   pdf_header("%PDF-2.0\r%なななな\n ".as_bytes()));
+    }
+
+    #[test]
+    fn integers_test() {
+        assert_eq!(PdfObject::Integer(123), signed_integer(b"123").to_result().unwrap());
+        assert_eq!(PdfObject::Integer(-123), signed_integer(b"-123").to_result().unwrap());
+        assert_eq!(PdfObject::Integer(0), signed_integer(b"0").to_result().unwrap());
+        assert_eq!(PdfObject::Integer(0), signed_integer(b"-0").to_result().unwrap()); // heh
+    }
+    macro_rules! float_test {
+        ($($name:ident: $value:expr,)*) => {
+        $(
+            #[test]
+            fn $name() {
+                let (input, expected) = $value;
+                match signed_float(input).to_result().unwrap() {
+                    PdfObject::Float(v) => {
+                        assert_relative_eq!(expected, v);
+                    },
+                    _ => {
+                        assert_eq!(5, 0);
+                    }
+                }
+            }
+        )*
+        }
+    }
+
+    float_test! {
+        f1: (b"123.0",123.0),
+        f2: (b"34.5",34.5),
+        f3: (b"-3.62",-3.62),
+        f4: (b"+123.6",123.6),
+        f5: (b"4.",4.0),
+        f6: (b"-.002",-0.002),
+        f7: (b"0.0",0.0),
+    }
+
+
+    #[test]
+    fn hexadecimal_string_test() {
+        assert_eq!(
+            nom::Err::Position(nom::ErrorKind::Tag, &[45u8, 62u8][..]),
+            hexadecimal_string(b"<a->".as_bytes()).to_result().unwrap_err()
+        );
+    }
+    macro_rules! hexst {
+        ($($name:ident: $value:expr,)*) => {
+            $(
+                #[test]
+                fn $name() {
+                    let (input, expected) = $value;
+                    match hexadecimal_string(input).to_result().unwrap() {
+                        PdfObject::String(v) => {
+                            assert_eq!( v, expected );
+                        },
+                        _ => {
+                            assert_eq!(7,0);
+                        }
+                    }
+                }
+            )*
+        }
+    }
+    hexst! {
+        hx1: (b"<abcdef0123456789>", vec![0xab, 0xcd, 0xef, 0x01, 0x23, 0x45, 0x67, 0x89]),
+        hx2: (b"<abc def0123\n456789 >", vec![0xab, 0xcd, 0xef, 0x01, 0x23, 0x45, 0x67, 0x89]),
+        hx3: (b"<ab>", vec![0xab]),
+        hx3a: (b"<a\rb>", vec![0xab]),
+        hx4: (b"<a>", vec![0xa0]),
+        hx5: (b"<ab cd\nef1>", vec![0xab, 0xcd, 0xef, 0x10]),
+    }
+
+    macro_rules! lsit {
+        ($($name:ident: $value:expr,)*) => {
+            $(
+                #[test]
+                fn $name() {
+                    let (input, expected) = $value;
+                    assert_eq!(expected,
+                        literal_string(input.as_bytes()).unwrap_inc());
+                }
+            )*
+        }
+    }
+
+
+    macro_rules! lset {
+        ($($name:ident: $value:expr,)*) => {
+            $(
+                #[test]
+                fn $name() {
+                    let (input, expected) = $value;
+                    assert_eq!(expected,
+                        literal_string(input.as_bytes()));
+                }
+            )*
+        }
+    }
+
+    lsit! {
+        lsit_1: (b"(abcde", Needed::Unknown),
+        lsit_2: (b"(abc()", Needed::Unknown),
+        lsit_3: (b"(abc\\", Needed::Unknown),
+    }
+
+    lset! {
+        lset_1: (b"(abc\\80)", Error(nom::Err::Position(ErrorKind::Custom(3333), b"(abc\\80)".as_bytes()))),
+    }
+
+    macro_rules! tlsr {
+        ($($name:ident: $value:expr,)*) => {
+            $(
+                #[test]
+                fn $name() {
+                    let (input, expected) = $value;
+                    match literal_string(input.as_bytes()).to_result().unwrap() {
+                        PdfObject::String(v) => {
+                            assert_eq!(expected[..].to_owned(), v);
+                        },
+                        _ => {
+                            assert_eq!(9,0);
+                        }
+                    }
+                }
+            )*
+        }
+    }
+
+    tlsr! {
+        tlsr_0: (b"(abcd)", b"abcd"),
+        tlsr_1: (b"(\\247)", b"\xA7"),
+
+        tlsr_2: (b"(a)", b"a"),
+        tlsr_3: (b"(This is a string)", b"This is a string"),
+        tlsr_4: (b"(Strings can contain newlines\nand such.)", b"Strings can contain newlines\nand such."),
+        tlsr_5: (b"(Strings can contain balanced parentheses ()\nand special characters ( * ! & } ^ %and so on) .)",
+                    b"Strings can contain balanced parentheses ()\nand special characters ( * ! & } ^ %and so on) ."),
+        tlsr_6: (b"(The following is an empty string .)", b"The following is an empty string ."),
+        tlsr_7: (b"()", b""),
+        tlsr_8: (b"(It has zero (0) length.)", b"It has zero (0) length."),
+
+        tlsr_9: (b"(These \\\rtwo strings \\\nare the same.)", b"These two strings are the same."),
+        tlsr_a: (b"(These two strings are the same.)", b"These two strings are the same."),
+
+        tlsr_b: (b"(This string has an end-of-line at the end of it.\n)", b"This string has an end-of-line at the end of it.\n"),
+        tlsr_c: (b"(So does this one.\\n)", b"So does this one.\n"),
+        tlsr_d: (b"(This string contains \\245two octal characters\\307.)", b"This string contains \xA5two octal characters\xC7."),
+        tlsr_e: (b"(\\0053)", b"\x053"),
+        tlsr_f: (b"(\\053)", b"+"),
+        tlsr_g: (b"(\\53)", b"+"),
+        tlsr_h: (b"(\\533)", b"+3"),
+    }
+
+    macro_rules! name_object_result_test {
+        ($($name:ident: $value:expr,)*) => {
+            $(
+                #[test]
+                fn $name() {
+                    let (input, expected) = $value;
+                    match name_object(input.as_bytes()).to_result().unwrap() {
+                        PdfObject::Name( v ) => {
+                            assert_eq!(expected[..].to_owned(), v);
+                        },
+                        _ => {
+                            assert_eq!(10, 0);
+                        }
+                    }
+                }
+            )*
+        }
+    }
+
+    name_object_result_test! {
+        nort_1: (b"/Name1",b"Name1"),
+        nort_2: (b"/ASomewhatLongerName",b"ASomewhatLongerName"),
+        nort_3:(b"/A;Name_With-Various***Characters?",b"A;Name_With-Various***Characters?"),
+        nort_5: (b"/1.2",b"1.2"),
+        nort_6: (b"/$$",b"$$"),
+        nort_7: (b"/@pattern",b"@pattern"),
+        nort_8: (b"/.notdef",b".notdef"),
+        nort_9: (b"/Lime#20Green",b"Lime Green"),
+        nort_10: (b"/paired#28#29parentheses",b"paired()parentheses"),
+        nort_11: (b"/The_Key_of_F#23_Minor",b"The_Key_of_F#_Minor"),
+        nort_12: (b"/A#42",b"AB"),
+        nort_13: (b"/#2F",b"/"),
+        nort_14: (b"/",b""),
+        nort_15: (b"/abcd[",b"abcd"),
+    }
+
+    #[test]
+    fn test_ws_recognize() {
+        assert_eq!(0, recognize_some_ws(b"a ".as_bytes()).to_result().unwrap().len());
+        assert_eq!(0, recognize_some_ws(b"".as_bytes()).to_result().unwrap().len());
+        assert_eq!(2, recognize_some_ws(b" \t".as_bytes()).to_result().unwrap().len());
+        assert_eq!(9, recognize_some_ws(b" \t \r\n \n \r".as_bytes()).to_result().unwrap().len());
+    }
+}
